+++ package.json
@@ -1,10 +1,14 @@
 {
   "name": "car-parts-backend",
   "version": "1.0.0",
-  "main": "index.js",
+  "type": "module",
+  "main": "index.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "run": "node index.ts"
+  },
+  "compilerOptions": {
+    "module": "CommonJS"
   },
   "keywords": [],
   "author": "",
@@ -16,6 +20,7 @@
   "dependencies": {
     "express": "^4.18.2",
     "mysql2": "^3.6.2",
+    "nodemon": "^3.0.1",
     "reflect-metadata": "^0.1.13",
     "typeorm": "^0.3.17"
   }
