+++ tsconfig.json
@@ -1,10 +1,9 @@
 {
    "compilerOptions": {
       "lib": [
-         "es5",
          "es6"
       ],
-      "target": "es5",
+      "target": "es6",
       "module": "commonjs",
       "moduleResolution": "node",
       "outDir": "./build",
